Make schemas consistent between NPM package and chronos
