+++ electron/models/DockerModel.ts
@@ -66,12 +66,4 @@ const DockerSchema = new Schema({
   },
 });
 
-// module.exports = service => mongoose.model(service, DockerSchema);
-
-//  const docker = (serviceName: any) => mongoose.model<any>(serviceName, DockerSchema);
-const docker = mongoose.model<IDock>('containerinfos', DockerSchema);
-// module.exports = (serviceName: any) => mongoose.model(serviceName, DockerSchema);
-export default docker;
-// export default mongoose.model<IDock>('docker', DockerSchema);
-// const DockerModelFunc = mongoose.model<IDock>('docker', DockerSchema)
-// export default DockerModelFunc;
+export default mongoose.model<IDock>('containerinfos', DockerSchema);
